lbs-ctl: refactor system and device information
Add --info and --device parameters to control wheter or not device (and
system) information is printed and for which device.

Output will now be shown in 3 slots: system, block layer and block
device information. If debug is passed, then other relevant lbs fields
are also added. Without debug, only the most important fields are shown.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>